move translation to its own folder WIP
